refactor NodeVisitor to Rectors
