+++ tests/Rector/Contrib/Nette/RemoveConfiguratorConstantsRector/Test.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace Rector\Tests\NodeVisitor\UpgradeDeprecation\ReplaceDeprecatedConstantNodeVisitor;
+namespace Rector\Tests\Rector\Contrib\Nette\RemoveConfiguratorConstantsRector;
 
 use Rector\Testing\PHPUnit\AbstractReconstructorTestCase;
