+++ tests/Rector/Contrib/Nette/RemoveConfiguratorConstantsRector/correct/correct.php.inc
