+++ tests/Rector/Contrib/Nette/RemoveConfiguratorConstantsRector/wrong/wrong.php.inc
